+++ docs/changelog.rst
@@ -9,6 +9,22 @@ The format is based on
 `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`__, and this project
 adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`__.
 
+Version 1.4.0 (2022-08-19)
+--------------------------
+
+Added
+*****
+
+- :meth:`Client.mark_notifications_as_seen` and
+  :meth:`Client.mark_notifications_as_read`.
+- :meth:`Notification.mark_as_seen` and :meth:`Notification.mark_as_read`.
+
+Removed
+*******
+
+- Removed support for python 3.6 as it has reached its end of life. For more
+  information, see `PEP 494 <https://peps.python.org/pep-0494/#lifespan>`__.
+
 Version 1.3.0 (2022-06-21)
 --------------------------